Numerous text/font related changes
Keep a flag to detect if we are running a CharProc, and another one to
detect if a CharProc has executed a d1 operation. If we encounter a
colour change in a CharProc which has executed a d1, we are required
to ignore the operation.

Have the read_typeX_font routines return the created gs_font structure
pointer, and gs_setfont from the pdfi_Tf() procedure to avoid having to
do all the tedious book-keeping in multiple places. Use a new function
pdfi_gs_setfont() instead of gs_setfont() and have that routine do the

Change the Widths array storage from double to float.

Check whether we are inside a text block when performing text operations
and NOT inside a text block when performing non-text marking operations;
raise an error flag if either occurs.

Check for nested text blocks, raise an error

Rework Tm and Tj. For now we will do exactly what the spec says and
maintain the CTM and text matrix separately, concatenating the two
when drawing text. Attempts to be clever and maintain the CTM fall down
in the face of the numerous malformed PDF files in the test suite.

There may be a better way to do this, but for now, this will do.

+++ pdf/pdf_font_types.h
@@ -89,7 +89,7 @@ typedef enum pdf_font_type_e {
     pdf_obj *Name;                  /* Should be a name object, but best allow for strings as well */\
     unsigned int FirstChar;         /* For PDF up to 1.4 this may be absent for the base 14 */\
     unsigned int LastChar;          /* For PDF up to 1.4 this may be absent for the base 14 */\
-    float *Widths;                  /* For PDF up to 1.4 this may be absent for the base 14 */\
+    double *Widths;                  /* For PDF up to 1.4 this may be absent for the base 14 */\
     pdf_array *Encoding;            /* Array built from name or dictionary */\
     pdf_obj *ToUnicode              /* Name or stream (technically should be a stream, but we've seen Identity names */
